Fix prepare images, now array of image corresponds to labels

+++ prepare_images.py
@@ -29,50 +29,52 @@ def compute_covering(window, label):
         return 0
 
 
-def split_into_subimgs(img, lbl, sub_img_shape, lbl_array, step=1):
+def split_into_subimgs(img, lbl, sub_img_shape, lbl_array, debug, step=1):
     shape = (lbl_array.shape[0], SUB_IMG_LAYERS, SUB_IMG_HEIGHT, SUB_IMG_WIDTH)
     result_array = as_strided(img, shape=shape,
                               strides=(img.strides[2] * step, img.strides[0], img.strides[1], img.strides[2]))
     index = 0
-    for i in range(0, img.shape[WIDTH] - sub_img_shape[WIDTH], step):
-        for ii in range(0, img.shape[HEIGHT] - sub_img_shape[HEIGHT], step):
-            # result_array.append(img[:, i:i + sub_img_shape[WIDTH], ii:ii + sub_img_shape[HEIGHT]])
+    for i in range(0, img.shape[HEIGHT] - sub_img_shape[HEIGHT], step):
+        for ii in range(0, img.shape[WIDTH] - sub_img_shape[WIDTH], step):
             # Rectangle = namedtuple('Rectangle', ['xmin', 'ymin', 'xmax', 'ymax'])
             is_cover = int(
-                compute_covering(Rectangle(i, ii, i + sub_img_shape[WIDTH], ii + sub_img_shape[HEIGHT]),
+                compute_covering(Rectangle(i, ii, i + sub_img_shape[HEIGHT], ii + sub_img_shape[WIDTH]),
                                  Rectangle(lbl[1], lbl[0], lbl[3], lbl[2])) > COVER_PERCENT)
-            if 1 == is_cover:
-                roi = img[:, i: i + sub_img_shape[WIDTH], ii: ii + sub_img_shape[HEIGHT]]
+            if debug and (1 == is_cover):
+                roi = img[:, i: i + sub_img_shape[HEIGHT], ii: ii + sub_img_shape[WIDTH]]
                 (r, g, b) = (roi[0], roi[1], roi[2])
                 roi = cv2.merge((r, g, b))
-                cv2.imshow("img", roi)
+                cv2.imshow("from labels", roi)
                 cv2.waitKey(0)
                 cv2.destroyAllWindows()
 
                 roi = result_array[index]
                 (r, g, b) = (roi[0], roi[1], roi[2])
                 roi = cv2.merge((r, g, b))
-                cv2.imshow("img", roi)
+                cv2.imshow("from strides", roi)
                 cv2.waitKey(0)
                 cv2.destroyAllWindows()
 
             lbl_array[index] = is_cover
             index += 1
+        index += sub_img_shape[WIDTH] / step
     return result_array
 
 
-def prepare(img_path, lbl):
+def prepare(img_path, lbl, debug=False):
     print("Prepare image " + img_path)
     print(lbl)
     step = 2
     img = cv2.imread(img_path, cv2.IMREAD_UNCHANGED)
+    print(img.shape)
     res_img = cv2.normalize(img.astype('float'), None, 0.0, 1.0, cv2.NORM_MINMAX)
     res_img = np.array([res_img[:, :, 0], res_img[:, :, 1], res_img[:, :, 2]])
 
     lbl_res = np.zeros(shape=int(
-        (res_img.shape[WIDTH] - SUB_IMG_WIDTH + 1) / step * (res_img.shape[HEIGHT] - SUB_IMG_HEIGHT + 1) / step))
+        (res_img.shape[WIDTH] - SUB_IMG_WIDTH + 1) / step * (res_img.shape[HEIGHT] - SUB_IMG_HEIGHT + 1) / step + (
+            SUB_IMG_WIDTH / step * res_img.shape[HEIGHT] / step)))
     res = split_into_subimgs(res_img, sub_img_shape=(SUB_IMG_LAYERS, SUB_IMG_HEIGHT, SUB_IMG_WIDTH), lbl=lbl,
-                             lbl_array=lbl_res, step=step)
+                             lbl_array=lbl_res, step=step, debug=debug)
 
     return res, lbl_res
